moving components to underscore folder so it's ignored by the router: _components
